pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "falcon-epdb"
version = "1.0.1"
description = ""
license = "BSD-3-Clause"
authors = [
    "Josh Wilson <josh.wilson@fivestars.com>"
]
readme = "README.rst"
homepage = "https://github.com/jcwilson/falcon-epdb"
repository = "https://github.com/jcwilson/falcon-epdb"
documentation = "https://falcon-epdb.readthedocs.io/"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 2",
    "Programming Language :: Python :: 2.7",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.4",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: CGI Tools/Libraries",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Middleware"
]

[tool.poetry.dependencies]
python = "~2.7 || ^3.4"
epdb = "^0.15.1"
cryptography = {version = "^2.5",optional = true}
PyJWT = {version = "^1.7",optional = true}

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pytest-mock = "^1.10"
sphinx = "^1.8"
sphinx_rtd_theme = "^0.4.2"
falcon = "^1.4"
pytest-sugar = "^0.9.2"
pytest-cov = "^2.6"
pydocstyle = "^3.0"
flake8 = "^3.7"
testfixtures = "^6.5"

[tool.poetry.extras]
fernet = ["cryptography"]
jwt = ["PyJWT"]

[tool.black]
line-length = 100
